TST: Fix some test builds for numpy 2.0 (#59046)
* TST: Fix some test builds for numpy 2.0

* 64 not 32

* Adjust some tests

* Revert "Adjust some tests"

This reverts commit ca28f250be92757fe3053019ff183139bb53fc56.

* Just pin numpy in pyarrow nightly build

* Mark test as pa under 17

+++ pandas/tests/scalar/timedelta/test_arithmetic.py
@@ -419,7 +419,7 @@ class TestTimedeltaMultiplicationDivision:
 
     def test_td_mul_numeric_ndarray_0d(self):
         td = Timedelta("1 day")
-        other = np.array(2)
+        other = np.array(2, dtype=np.int64)
         assert other.ndim == 0
         expected = Timedelta("2 days")
